refactor(utxo-list): externalize UtxoDetailModal
