+++ src/components/jar_details/JarDetailsOverlay.module.css
@@ -1,38 +1,3 @@
-.utxoDetailModalBackdrop {
-  z-index: 1300 !important;
-}
-
-.utxoDetailModal {
-  z-index: 1400 !important;
-}
-
-.modalHeader {
-  display: flex !important;
-  justify-content: flex-start !important;
-  background-color: transparent !important;
-  padding: 1.25rem !important;
-}
-
-.modalTitle {
-  width: 100%;
-  font-size: 1rem !important;
-  font-weight: 400 !important;
-  color: var(--bs-body-color) !important;
-}
-
-.modalTitle > div:first-child {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.modalTitle .cancelButton {
-  padding: 0 0 0 1rem;
-  color: var(--bs-body-color);
-  background-color: transparent !important;
-  border: none;
-}
-
 .overlayContainer .accountStepperTitle {
   display: inline-flex;
   justify-content: center;
